further functionality for api calls function and integration, with some styling
