+++ package.json
@@ -12,6 +12,7 @@
     "@types/node": "^16.18.105",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
+    "axios": "^1.7.5",
     "emailjs-com": "^3.2.0",
     "lucide-react": "^0.436.0",
     "notistack": "^3.0.1",
