+++ src/Providers-page/ProvidersPage.css
@@ -113,7 +113,8 @@
 
 .view-details-button,
 .view-on-map-button,
-.view-on-map-dropdown {
+.view-on-map-dropdown,
+.pagination-button {
     font-size: 1.125em;
     font-weight: 600;
     line-height: 1.2;
@@ -157,8 +158,6 @@
     text-align: left;
     margin-bottom: 0.5rem;
 }
-
-
 
 .searched-provider-card:hover {
     background: #c4b56b;
@@ -206,10 +205,14 @@
     cursor: pointer;
     margin: 0 5px;
     border-radius: 5px;
+    box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.2);
+    transition: background-color 0.3s, transform 0.3s; 
 }
 
 .pagination-button:hover {
-    background-color: #0056b3;
+    background-color: white;
+    color: #007bff;
+    transform: translateY(-5px);
 }
 
 .view-on-map-dropdown {
@@ -220,7 +223,6 @@
     font-size: 1.3rem;
     cursor: pointer;
     padding: 0.5rem;
-    border-radius: 0.25em;
     box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.2);
     transition: background-color 0.3s, transform 0.3s;
     text-align: left;
