+++ src/Utility/Types.ts
@@ -1,19 +1,51 @@
-export interface Provider {
-    name: string;
-    website: string;
-    phone: string;
-    email: string;
-    insurances: string;
-    locations: string;
-    rates: string;
-    ages: string;
-    waitlist: string;
-    telehealth: string;
-    spanish: string;
+export interface Insurance {
+    name: string | null;
 }
 
-export interface Screening {
-    question: string;
-    answer: string;
+export interface Location {
+    name: string | null;
+    address_1: string | null;
+    address_2: string | null;
+    city: string | null;
+    state: string | null;
+    zip: string | null;
+    phone: string | null;
 }
 
+export interface CountiesServed {
+    county: string | null;
+}
+
+export interface Services {
+    telehealth_services: string | null;
+    spanish_speakers: string | null;
+    at_home_services: string | null;
+    in_clinic_services: string | null;
+}
+
+export interface ProviderAttributes {
+    name: string | null;
+    locations: Location[];
+    insurance: Insurance[];
+    counties_served: CountiesServed[];
+    website: string | null;
+    email: string | null;
+    cost: string | null;
+    min_age: number | null;
+    max_age: number | null;
+    waitlist: string | null;
+    telehealth_services: string | null;
+    spanish_speakers: string | null;
+    at_home_services: string | null;
+    in_clinic_services: string | null;
+}
+
+export interface MockProviderData {
+    id: number;
+    type: string;
+    attributes: ProviderAttributes;
+}
+
+export interface MockProviders {
+    data: MockProviderData[];
+}
